backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "AntonKochurka",email = "antonkotchurka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.121.1,<0.122.0)",
    "sqlalchemy[async] (>=2.0.44,<3.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
black = "^25.11.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"

